refact: 결제 정보 삭제 api PaymentAdminController로 이동
