+++ api-server/src/main/java/com/ThreeK_Project/api_server/domain/payment/controller/PaymentAdminController.java
@@ -3,13 +3,15 @@ package com.ThreeK_Project.api_server.domain.payment.controller;
 import com.ThreeK_Project.api_server.domain.payment.dto.PaymentResponseDto;
 import com.ThreeK_Project.api_server.domain.payment.dto.PaymentSearchDto;
 import com.ThreeK_Project.api_server.domain.payment.service.PaymentService;
+import com.ThreeK_Project.api_server.global.dto.SuccessResponse;
+import com.ThreeK_Project.api_server.global.security.auth.UserDetailsCustom;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.ModelAttribute;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.web.bind.annotation.*;
+
+import java.util.UUID;
 
 @RestController
 @RequiredArgsConstructor
@@ -21,5 +23,12 @@ public class PaymentAdminController {
     @GetMapping
     public ResponseEntity<Page<PaymentResponseDto>> searchPayments(@ModelAttribute PaymentSearchDto searchDto) {
         return ResponseEntity.ok(paymentService.searchPayments(searchDto));
+    }
+
+    @DeleteMapping("/{paymentId}")
+    public ResponseEntity<SuccessResponse> deletePayment(@PathVariable("paymentId") UUID paymentId) {
+        UserDetailsCustom userDetails = (UserDetailsCustom) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        paymentService.deletePayment(paymentId, userDetails.getUser());
+        return ResponseEntity.ok(new SuccessResponse("결제 정보 삭제 성공"));
     }
 }
